requirements-dev.txt
pytest>=8.0
pytest-asyncio>=0.23
pytest-homeassistant-custom-component>=0.13
pytest-cov>=5.0
black>=24.0
